email_signals/requirements.txt
django-ckeditor==6.1.0
